+++ lib/src/extensions/extensions.dart
@@ -1,6 +1,6 @@
 export 'src/datetime_extensions.dart';
 export 'src/getx_extensions.dart';
 export 'src/iterable_extensions.dart';
+export 'src/misc_extensions.dart';
 export 'src/num_extensions.dart';
 export 'src/string_extensions.dart';
-export 'src/misc_extensions.dart';